mini-rag-app/rag-app/requirements.txt
streamlit>=1.30.0
faiss-cpu>=1.7.4
openai>=1.10.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
numpy>=1.24.0
